Add nightly-publish workflow and game settings for combat tweaks
This commit adds a new workflow `nightly-publish.yml` that runs on a schedule and manually. It checks for changes in the repository using Git log and outputs whether it should run or not. If there are changes, it proceeds to publish the mod files. The publishing process includes fetching Spriggit, extracting it, fixing permissions, running Spriggit to deserialize YAML files, deleting unnecessary files, zipping the mod files, deleting existing releases with the tag "nightly", and creating a new release with the tag "nightly" containing the zipped mod files.

Additionally, this commit introduces new game settings for combat tweaks in Spriggit. These include `fCombatDodgeReactionTime`, `fCombatEvadeReactionTime`, `fCombatMeleeIdleTime`, `fCombatRangedGrenadeIdleTime`, `fCombatRangedGunAnimationDrivenDelayTime`, and `fCombatRangedGunFireWaitTimeMax`. These settings are defined in YAML format under the directory `/spriggit/GameSettings/`.

The README file has also been updated to change the title from "Stealth-and-Combat-Tweaks" to "Stealth and Combat Tweaks".